refactor: découpe TableauView en sous-composants
TableauScoreModal, TableauPendingSection, TableauToolbar extraits.
TableauView réduit significativement.
